abstracting code
